# Optimize TaskItem cloning between Engine and Tasks