--- a/src/Build/Instance/ImmutableProjectCollections/ImmutablePropertyCollectionConverter.cs
+++ b/src/Build/Instance/ImmutableProjectCollections/ImmutablePropertyCollectionConverter.cs
@@ -68,5 +68,7 @@ namespace Microsoft.Build.Instance
 
             return true;
         }
+
+        public bool HasSameBacking(ICopyOnWritePropertyDictionary<T> other) => ReferenceEquals(other, this);
     }
 }