--- a/src/Framework/MSBuildNameIgnoreCaseComparer.cs
+++ b/src/Framework/MSBuildNameIgnoreCaseComparer.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections.Generic;
 using Microsoft.Build.Framework;
-using Microsoft.Build.Shared;
 
 #nullable disable