… src/Build/Microsoft.Build.csproj
… src/Framework/CopyOnWriteDictionary.cs
… src/Framework/IConstrainedEqualityComparer.cs
… src/Framework/IMetadataContainer.cs
… src/Framework/TaskItemData.cs
… src/MSBuild/MSBuild.csproj
… src/MSBuildTaskHost/MSBuildTaskHost.csproj
… src/Shared/TaskParameter.cs
… src/Tasks/AssemblyDependency/ReferenceTable.cs
… src/Tasks/Microsoft.Build.Tasks.csproj
… src/Utilities/Microsoft.Build.Utilities.csproj
… src/Utilities/TaskItem.cs