### Context

Currently the engine and task implementations of `ITaskItem` will always do a full rebuild of their backing `ImmutableDictionary<string, T>`. This is due to having different backing types (string vs `ProjectMetadataInstance`), plus `CopyOnWriteDictionary<T>` type living under `Shared`, preventing type guards from working even if it *did* expose the bakcing dictionary.

This affects both Tasks copying metadata within a task, as well as the engine whenever any task completes execution, as `Utilities.TaskItem` must always be converted back into `ProjectItemInstance.TaskItem`. See `GatherTaskItemOutputs` in a profiler to see how expensive this is, with most of the impact coming from RAR.

### Changes Made

This PR is largely two parts:
- Moving types around and modifying the backing collections which use `ImmutableDictionary` to allow reference equality checks.
- Giving `TaskITem` implementations the ability to clone from other implementations when given a `CopyOnWriteDictionary` instance.

Also adds a simple optimization in RAR to cache any `ProjectItemInstance.TaskItem` instances which came from a source file into temporary `Utilities.TaskItem`, avoiding the expensive enumeration. These are reused across references within each RAR invocation, and the shared `ImmutableDictionary` types allow us to do this for free. (Could probably split that off into its own PR to wait till this is merged, but it's a couple lines and net negative perf without this so...)

### Testing

Example of spending 1718ms in `GatherTaskItemOutputs` alone:
![image](https://github.com/user-attachments/assets/2dc632c8-2057-48d1-95b5-a3fdfa99f451)

374 ms with this change. Note that the `TaskItem` copying doesn't even show up in the profile here since it's just a no-op clone, so it's effectively free and the rest is unrelated overhead.
![image](https://github.com/user-attachments/assets/53bc2aba-8b15-4b30-ab69-efcf9f81f3a5)

Need to test on more repos, and other hotspots are hard to track since these get passed around everywhere - but there's a pretty clear difference just running ad-hoc builds back-to-back:

This PR:
```
     2371 ms  MSBuild                                  1053 calls
     8232 ms  ResolveAssemblyReference                 142 calls
...
Time Elapsed 00:00:38.69
```
Mainline:
```
     2537 ms  MSBuild                                  1053 calls
     9117 ms  ResolveAssemblyReference                 142 calls
...
Time Elapsed 00:00:40.73
```


### Notes
Pretty confident on the overall design, but still need to UT and see if there are places consolidate code since there's a fair amount of type casting and duplication in the TaskItem CopyTo() / ImportMetadata() methods that I think can be cleaned up still. May be able split this into A. Moves / modyfing the types + B. Optimized cloning