@ccastanedaucf 2025-03-28
The failing tests all fail when I checked out upstream, so not sure what's going on there. Otherwise I've had great numbers building on other repos.

I just tested a few runs on Orchard Core and it's around 37% cut from RAR alone, plus seems like every task is consistently faster.

Multi-proc:
`...\artifacts\bin\bootstrap\core\dotnet.exe build -tl:off -restore:false /clp:PerformanceSummary /m`

Before:
```
      891 ms  Hash                                     1064 calls
      974 ms  Microsoft.Build.Tasks.Git.GetUntrackedFiles 420 calls
     1116 ms  AssignProjectConfiguration               436 calls
     1243 ms  ResolvePackageAssets                     436 calls
     1414 ms  ResolveTargetingPackAssets               436 calls
     3039 ms  Copy                                     7037 calls
     5140 ms  ResolvePackageFileConflicts              436 calls
    19162 ms  ResolveAssemblyReference                 436 calls
    132508 ms  CallTarget                               872 calls
    3940843 ms  MSBuild                                  2388 calls

Time Elapsed 00:00:26.55
```

After:
```
      696 ms  Hash                                     1064 calls
      776 ms  ResolvePackageAssets                     436 calls
      814 ms  Microsoft.Build.Tasks.Git.GetUntrackedFiles 420 calls
      869 ms  ProcessFrameworkReferences               436 calls
      890 ms  AssignProjectConfiguration               436 calls
     2636 ms  Copy                                     7037 calls
     4720 ms  ResolvePackageFileConflicts              436 calls
    12221 ms  ResolveAssemblyReference                 436 calls
    82668 ms  CallTarget                               872 calls
    4317227 ms  MSBuild                                  2388 calls

Time Elapsed 00:00:24.46
```

Single proc (still important for quickbuild):
`...\artifacts\bin\bootstrap\core\dotnet.exe build -tl:off -restore:false /clp:PerformanceSummary /m:1`

Before
```
      385 ms  Hash                                     1064 calls
      457 ms  ProcessFrameworkReferences               436 calls
      530 ms  AssignProjectConfiguration               436 calls
      531 ms  Microsoft.Build.Tasks.Git.GetUntrackedFiles 420 calls
      770 ms  ResolvePackageAssets                     436 calls
      781 ms  ResolveTargetingPackAssets               436 calls
     1165 ms  CallTarget                               872 calls
     2122 ms  Copy                                     7037 calls
     2993 ms  ResolvePackageFileConflicts              436 calls
    10429 ms  ResolveAssemblyReference                 436 calls
    321574 ms  MSBuild                                  2388 calls

Time Elapsed 00:01:12.34
```
After
```
      362 ms  Hash                                     1064 calls
      405 ms  ProcessFrameworkReferences               436 calls
      436 ms  AssignProjectConfiguration               436 calls
      504 ms  Microsoft.Build.Tasks.Git.GetUntrackedFiles 420 calls
      651 ms  ResolvePackageAssets                     436 calls
      751 ms  WriteLinesToFile                         1492 calls
     1016 ms  CallTarget                               872 calls
     1887 ms  Copy                                     7037 calls
     2874 ms  ResolvePackageFileConflicts              436 calls
     6522 ms  ResolveAssemblyReference                 436 calls
    286337 ms  MSBuild                                  2388 calls

Time Elapsed 00:01:04.97
```